requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
stripe==10.11.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
